+++ IBJts/samples/Java/samples/rfq/SampleRfq.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2019 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
+/* Copyright (C) 2024 Interactive Brokers LLC. All rights reserved. This code is subject to the terms
  * and conditions of the IB API Non-Commercial License or the IB API Commercial License, as applicable. */
 
 package samples.rfq;
@@ -350,7 +350,7 @@ public class SampleRfq extends SimpleWrapper {
       }
    }
    public void orderStatus(int orderId, String status, int filled,
-         int remaining, double avgFillPrice, int permId, int parentId,
+         int remaining, double avgFillPrice, long permId, int parentId,
          double lastFillPrice, int clientId, String whyHeld) {
       consoleMsg("orderStatus:" + orderId + " status=" + status);
